Add files via upload
Problem Statement:
The Agriculture business domain, as a vital part of the overall supply chain, is expected
to highly evolve in the upcoming years via the developments, which are taking place on
the side of the Future Internet. This paper presents a novel Business-to-Business
collaboration platform from the agri-food sector perspective, which aims to facilitate the
collaboration of numerous stakeholders belonging to associated business domains, in an
effective and flexible manner.